requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
Django==4.2.2
django-click==2.3.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
python-decouple==3.8
sqlparse==0.4.4
